refactor: move tables components into components/tables
