+++ components/tables/transactions-table.tsx
@@ -6,7 +6,7 @@ import {
    TableHead,
    TableHeader,
    TableRow,
-} from "./ui/table";
+} from "../ui/table";
 
 export default function TransactionsTable() {
    return (
